// client/src/generated/walrus_rut_state_type.ts
// THIS FILE IS AUTOMATICALLY GENERATED BY SPACETIMEDB. EDITS TO THIS FILE
// WILL NOT BE SAVED. MODIFY TABLES IN YOUR MODULE SOURCE CODE INSTEAD.

/* eslint-disable */
/* tslint:disable */
import {
  AlgebraicType as __AlgebraicTypeValue,
  BinaryReader as __BinaryReader,
  BinaryWriter as __BinaryWriter,
  ClientCache as __ClientCache,
  ConnectionId as __ConnectionId,
  DbConnectionBuilder as __DbConnectionBuilder,
  DbConnectionImpl as __DbConnectionImpl,
  Identity as __Identity,
  SubscriptionBuilderImpl as __SubscriptionBuilderImpl,
  TableCache as __TableCache,
  TimeDuration as __TimeDuration,
  Timestamp as __Timestamp,
  deepEqual as __deepEqual,
  type AlgebraicType as __AlgebraicTypeType,
  type AlgebraicTypeVariants as __AlgebraicTypeVariants,
  type CallReducerFlags as __CallReducerFlags,
  type ErrorContextInterface as __ErrorContextInterface,
  type Event as __Event,
  type EventContextInterface as __EventContextInterface,
  type ReducerEventContextInterface as __ReducerEventContextInterface,
  type SubscriptionEventContextInterface as __SubscriptionEventContextInterface,
  type TableHandle as __TableHandle,
} from "spacetimedb";

export type WalrusRutState = {
  id: number,
  currentCycleDay: number,
  isRutActive: boolean,
  lastUpdateTimestamp: __Timestamp,
};
let _cached_WalrusRutState_type_value: __AlgebraicTypeType | null = null;

/**
 * An object for generated helper functions.
 */
export const WalrusRutState = {
  /**
  * A function which returns this type represented as an AlgebraicType.
  * This function is derived from the AlgebraicType used to generate this type.
  */
  getTypeScriptAlgebraicType(): __AlgebraicTypeType {
    if (_cached_WalrusRutState_type_value) return _cached_WalrusRutState_type_value;
    _cached_WalrusRutState_type_value = __AlgebraicTypeValue.Product({ elements: [] });
    _cached_WalrusRutState_type_value.value.elements.push(
      { name: "id", algebraicType: __AlgebraicTypeValue.U32 },
      { name: "currentCycleDay", algebraicType: __AlgebraicTypeValue.U32 },
      { name: "isRutActive", algebraicType: __AlgebraicTypeValue.Bool },
      { name: "lastUpdateTimestamp", algebraicType: __AlgebraicTypeValue.createTimestampType() },
    );
    return _cached_WalrusRutState_type_value;
  },

  serialize(writer: __BinaryWriter, value: WalrusRutState): void {
    __AlgebraicTypeValue.serializeValue(writer, WalrusRutState.getTypeScriptAlgebraicType(), value);
  },

  deserialize(reader: __BinaryReader): WalrusRutState {
    return __AlgebraicTypeValue.deserializeValue(reader, WalrusRutState.getTypeScriptAlgebraicType());
  },

}

export default WalrusRutState;